# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Online Examination Proctoring System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegisterInputRole = typeof RegisterInputRole[keyof typeof RegisterInputRole];


export const RegisterInputRole = {
  student: 'student',
  proctor: 'proctor',
  admin: 'admin',
} as const;

export interface RegisterInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 6 */
  password: string;
  role: RegisterInputRole;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  student: 'student',
  proctor: 'proctor',
  admin: 'admin',
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  /** @nullable */
  avatarUrl?: string | null;
  isActive?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export type UserUpdateRole = typeof UserUpdateRole[keyof typeof UserUpdateRole];


export const UserUpdateRole = {
  student: 'student',
  proctor: 'proctor',
  admin: 'admin',
} as const;

export interface UserUpdate {
  name?: string;
  email?: string;
  role?: UserUpdateRole;
  isActive?: boolean;
}

export type ExamStatus = typeof ExamStatus[keyof typeof ExamStatus];


export const ExamStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface Exam {
  id: number;
  title: string;
  /** @nullable */
  description?: string | null;
  /** Duration in minutes */
  duration: number;
  /** @nullable */
  startTime?: string | null;
  /** @nullable */
  endTime?: string | null;
  status: ExamStatus;
  maxViolations?: number;
  /** @nullable */
  proctorId?: number | null;
  totalQuestions?: number;
  totalPoints?: number;
  passingScore?: number;
  createdAt: string;
}

export interface ExamInput {
  title: string;
  description?: string;
  duration: number;
  startTime?: string;
  endTime?: string;
  maxViolations?: number;
  proctorId?: number;
  passingScore?: number;
}

export type ExamUpdateStatus = typeof ExamUpdateStatus[keyof typeof ExamUpdateStatus];


export const ExamUpdateStatus = {
  draft: 'draft',
  scheduled: 'scheduled',
  active: 'active',
  completed: 'completed',
  cancelled: 'cancelled',
} as const;

export interface ExamUpdate {
  title?: string;
  description?: string;
  duration?: number;
  startTime?: string;
  endTime?: string;
  status?: ExamUpdateStatus;
  maxViolations?: number;
  proctorId?: number;
  passingScore?: number;
}

export interface ExamStats {
  examId: number;
  totalEnrolled: number;
  totalCompleted: number;
  totalFlagged: number;
  averageScore: number;
  passRate: number;
  violationCount: number;
}

export type QuestionType = typeof QuestionType[keyof typeof QuestionType];


export const QuestionType = {
  mcq: 'mcq',
  true_false: 'true_false',
  short_answer: 'short_answer',
} as const;

export interface Question {
  id: number;
  examId: number;
  type: QuestionType;
  content: string;
  /**
     * JSON string of options array for MCQ
     * @nullable
     */
  options?: string | null;
  /** @nullable */
  correctAnswer?: string | null;
  points: number;
  order: number;
}

export type QuestionInputType = typeof QuestionInputType[keyof typeof QuestionInputType];


export const QuestionInputType = {
  mcq: 'mcq',
  true_false: 'true_false',
  short_answer: 'short_answer',
} as const;

export interface QuestionInput {
  type: QuestionInputType;
  content: string;
  options?: string;
  correctAnswer?: string;
  points: number;
  order?: number;
}

export type QuestionUpdateType = typeof QuestionUpdateType[keyof typeof QuestionUpdateType];


export const QuestionUpdateType = {
  mcq: 'mcq',
  true_false: 'true_false',
  short_answer: 'short_answer',
} as const;

export interface QuestionUpdate {
  type?: QuestionUpdateType;
  content?: string;
  options?: string;
  correctAnswer?: string;
  points?: number;
  order?: number;
}

export type EnrollmentStatus = typeof EnrollmentStatus[keyof typeof EnrollmentStatus];


export const EnrollmentStatus = {
  enrolled: 'enrolled',
  active: 'active',
  submitted: 'submitted',
  flagged: 'flagged',
  disqualified: 'disqualified',
} as const;

export interface Enrollment {
  id: number;
  examId: number;
  userId: number;
  status: EnrollmentStatus;
  /** @nullable */
  startedAt?: string | null;
  /** @nullable */
  submittedAt?: string | null;
  violationCount?: number;
  exam?: Exam;
  user?: User;
  createdAt: string;
}

export interface Answer {
  id: number;
  enrollmentId: number;
  questionId: number;
  /** @nullable */
  answer?: string | null;
  /** @nullable */
  isCorrect?: boolean | null;
  score?: number;
  createdAt: string;
}

export interface AnswerInput {
  questionId: number;
  answer: string;
}

export type ViolationType = typeof ViolationType[keyof typeof ViolationType];


export const ViolationType = {
  face_missing: 'face_missing',
  multiple_faces: 'multiple_faces',
  eye_deviation: 'eye_deviation',
  audio_detected: 'audio_detected',
  tab_switch: 'tab_switch',
  fullscreen_exit: 'fullscreen_exit',
  copy_paste: 'copy_paste',
  phone_detected: 'phone_detected',
} as const;

export type ViolationSeverity = typeof ViolationSeverity[keyof typeof ViolationSeverity];


export const ViolationSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface Violation {
  id: number;
  enrollmentId: number;
  type: ViolationType;
  severity: ViolationSeverity;
  /** @nullable */
  details?: string | null;
  timestamp: string;
}

export type ViolationInputType = typeof ViolationInputType[keyof typeof ViolationInputType];


export const ViolationInputType = {
  face_missing: 'face_missing',
  multiple_faces: 'multiple_faces',
  eye_deviation: 'eye_deviation',
  audio_detected: 'audio_detected',
  tab_switch: 'tab_switch',
  fullscreen_exit: 'fullscreen_exit',
  copy_paste: 'copy_paste',
  phone_detected: 'phone_detected',
} as const;

export type ViolationInputSeverity = typeof ViolationInputSeverity[keyof typeof ViolationInputSeverity];


export const ViolationInputSeverity = {
  low: 'low',
  medium: 'medium',
  high: 'high',
  critical: 'critical',
} as const;

export interface ViolationInput {
  type: ViolationInputType;
  severity: ViolationInputSeverity;
  details?: string;
}

export interface ExamResult {
  id: number;
  enrollmentId: number;
  totalScore: number;
  maxScore: number;
  percentage: number;
  /** @nullable */
  grade?: string | null;
  passed: boolean;
  violationCount?: number;
  enrollment?: Enrollment;
  createdAt: string;
}

export interface DashboardStats {
  totalExams: number;
  activeExams: number;
  totalStudents: number;
  totalProctors: number;
  ongoingSessions: number;
  flaggedSessions: number;
  completedToday: number;
  averagePassRate: number;
}

export type ActivityItemType = typeof ActivityItemType[keyof typeof ActivityItemType];


export const ActivityItemType = {
  exam_started: 'exam_started',
  exam_submitted: 'exam_submitted',
  violation_detected: 'violation_detected',
  student_flagged: 'student_flagged',
  exam_created: 'exam_created',
} as const;

export interface ActivityItem {
  id: number;
  type: ActivityItemType;
  message: string;
  /** @nullable */
  studentName?: string | null;
  /** @nullable */
  examTitle?: string | null;
  timestamp: string;
}

export interface ViolationSummaryItem {
  type: string;
  count: number;
  percentage: number;
}

export type ListUsersParams = {
role?: string;
search?: string;
};

export type ListExamsParams = {
status?: string;
search?: string;
};