# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: AI Online Examination Proctoring System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/register:
    post:
      operationId: register
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Bad request
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /exams/{id}/enroll:
    post:
      operationId: enrollInExam
      tags: [enrollments]
      summary: Enroll in exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /exams/{id}/enrollments:
    get:
      operationId: listExamEnrollments
      tags: [enrollments]
      summary: List enrollments for an exam
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Enrollment"
...
  /enrollments/{id}:
    get:
      operationId: getEnrollment
      tags: [enrollments]
      summary: Get enrollment by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}/start:
    post:
      operationId: startExam
      tags: [enrollments]
      summary: Start the exam session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Enrollment"
  /enrollments/{id}/submit:
    post:
      operationId: submitExam
      tags: [enrollments]
      summary: Submit exam and end session
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExamResult"
  /enrollments/{id}/answers:
    get:
      operationId: listAnswers
      tags: [answers]
      summary: Get answers for enrollment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Answer"
    post:
      operationId: submitAnswer
      tags: [answers]
      summary: Submit or update an answer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Answer"
...
components:
  schemas:
...
    RegisterInput:
      type: object
      required: [name, email, password, role]
      properties:
        name:
          type: string
          minLength: 2
        email:
          type: string
        password:
          type: string
          minLength: 6
        role:
          type: string
          enum: [student, proctor, admin]
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [student, proctor, admin]
        avatarUrl:
          type: ["string", "null"]
        isActive:
          type: boolean
        createdAt:
          type: string
...
    Exam:
      type: object
      required: [id, title, duration, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        duration:
          type: integer
          description: Duration in minutes
        startTime:
          type: ["string", "null"]
        endTime:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, scheduled, active, completed, cancelled]
        maxViolations:
          type: integer
        proctorId:
          type: ["integer", "null"]
        totalQuestions:
          type: integer
        totalPoints:
          type: integer
        passingScore:
          type: integer
        createdAt:
          type: string
...
    Enrollment:
      type: object
      required: [id, examId, userId, status, createdAt]
      properties:
        id:
          type: integer
        examId:
          type: integer
        userId:
          type: integer
        status:
          type: string
          enum: [enrolled, active, submitted, flagged, disqualified]
        startedAt:
          type: ["string", "null"]
        submittedAt:
          type: ["string", "null"]
        violationCount:
          type: integer
        exam:
          $ref: "#/components/schemas/Exam"
        user:
          $ref: "#/components/schemas/User"
        createdAt:
          type: string
    Answer:
      type: object
      required: [id, enrollmentId, questionId, createdAt]
      properties:
        id:
          type: integer
        enrollmentId:
          type: integer
        questionId:
          type: integer
        answer:
          type: ["string", "null"]
        isCorrect:
          type: ["boolean", "null"]
        score:
          type: integer
        createdAt:
          type: string
    AnswerInput:
      type: object
      required: [questionId, answer]
      properties:
        questionId:
          type: integer
        answer:
          type: string
...
    ExamResult:
      type: object
      required: [id, enrollmentId, totalScore, maxScore, percentage, passed, createdAt]
      properties:
        id:
          type: integer
        enrollmentId:
          type: integer
        totalScore:
          type: integer
        maxScore:
          type: integer
        percentage:
          type: number
        grade:
          type: ["string", "null"]
        passed:
          type: boolean
        violationCount:
          type: integer
        enrollment:
          $ref: "#/components/schemas/Enrollment"
        createdAt:
          type: string
...